use new shared component
